enumflag: Document the order of flag keys.
For reasons lost to time, I was sorting the keys. This makes nice output (and
maybe that's why) but it makes it harder to reason about the correspondence
between keys and indices. Go back to the simpler behaviour where the default is
at 0 and everything else is wherever the caller put it.

This time, document it.